Chore: Merge pages and page permissions
